requirements.txt
streamlit_chat
streamlit_feedback
ibm-watson-machine-learning
ibm-cloud-sdk-core
crewai
python-dotenv
loguru
langchain-milvus
ibm-watsonx-ai
langchain_community
langchain-ibm
streamlit==1.31.1
streamlit-chat==0.1.1
streamlit-feedback==0.1.3
python-dotenv==1.0.1
loguru==0.7.2
ibm-watson-machine-learning==1.0.312
arize-phoenix
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-crewai
opentelemetry-exporter-otlp==1.23.0
openinference-instrumentation-litellm